Move connect() to new client module
